Refactor names from community to profile
