+++ src/components/Community/communityComponents/CommunityDropdown.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const CommunityDropdown = () => {
+  return (
+    <div className="w-full h-[165px]">
+      
+    </div>
+  )
+}
+
+export default CommunityDropdown
